refactor: Update import path for ModalFullScreen in LazyImg.vue
