+++ components/content/LazyImg.vue
@@ -2,7 +2,7 @@
 import { withTrailingSlash, withLeadingSlash, joinURL } from 'ufo';
 import { useRuntimeConfig, computed } from '#imports';
 import { ModalsContainer, useModal } from 'vue-final-modal';
-import ModalFullscreen from '~/components/ModalFullScreen.vue';
+import Modal from '~/components/FullScreenModal.vue';
 
 const props = defineProps({
 	src: {
@@ -45,7 +45,7 @@ const refinedSrc = computed(() => {
 	return props.src;
 });
 const { open, close } = useModal({
-	component: ModalFullscreen,
+	component: Modal,
 	attrs: {
 		onClose() {
 			close();
